Move backoff function
