+++ tests/manager/celery/tasks/test_search.py
@@ -5,7 +5,6 @@ from celery.exceptions import MaxRetriesExceededError
 from opensearchpy import OpenSearchException
 
 from palace.manager.celery.tasks.search import (
-    exponential_backoff,
     get_migrate_search_chain,
     get_work_search_documents,
     index_work,
@@ -18,26 +17,6 @@ from tests.fixtures.database import DatabaseTransactionFixture
 from tests.fixtures.search import EndToEndSearchFixture
 from tests.fixtures.services import ServicesFixture
 from tests.mocks.search import MockSearchSchemaRevisionLatest
-
-
-@pytest.mark.parametrize(
-    "retries, expected",
-    [
-        (0, 3),
-        (1, 9),
-        (2, 27),  # ~0.5 minutes
-        (3, 81),  # ~1.3 minutes
-        (4, 243),  # ~4 minutes
-        (5, 729),  # ~12 minutes
-    ],
-)
-def test_exponential_backoff(retries: int, expected: int) -> None:
-    with patch(
-        "palace.manager.celery.tasks.search.randrange", return_value=0
-    ) as mock_randrange:
-        assert exponential_backoff(retries) == expected
-    assert mock_randrange.call_count == 1
-    mock_randrange.assert_called_with(0, round(expected * 0.3))
 
 
 def test_get_work_search_documents(db: DatabaseTransactionFixture) -> None:
